Move constants and regex to their own files from index and fields file
